requirements.txt
numpy
pandas
tqdm
opencv-python
torch
torchvision
hydra-core==1.1.1
omegaconf
tensorboard
mmcv==1.3.11
einops==0.3.0
timm==0.4.12
gdown
matplotlib
scikit-learn
menovideo
jinja2
ipykernel
zenodo-get
gdown